Projects/sharry boii project/pyproject.toml
[project]
name = "sharryboii-live"
version = "0.1.0"
description = "sharryboii is a ai agent inspired by gemini-live"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elevenlabs>=2.4.0",
    "gradio>=5.34.2",
    "groq>=0.28.0",
    "gtts>=2.5.4",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.5",
    "langgraph>=0.4.8",
    "opencv-python>=4.11.0.86",
    "psutil>=7.0.0",
    "pyaudio>=0.2.14",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.0",
    "speechrecognition>=3.14.3",
]
